[TWINFACES-0] feat: update schema, fix: comments types, refactor: remove resource link in twinclass table row name
