+++ src/entities/comment/components/CommentCard.tsx
@@ -1,6 +1,6 @@
 import { CircleUserRound, EllipsisVertical } from "lucide-react";
 
-import { CommentView_DETAILED } from "@/entities/comment";
+import { Comment_DETAILED } from "@/entities/comment";
 import { formatToTwinfaceDate, isPopulatedArray } from "@/shared/libs";
 import {
   Avatar,
@@ -15,9 +15,9 @@ import {
 
 import { UserResourceLink } from "../../user";
 
-interface CommentCardProps {
-  item: CommentView_DETAILED;
-}
+type CommentCardProps = {
+  item: Comment_DETAILED;
+};
 
 export function CommentCard({ item }: CommentCardProps) {
   return (
@@ -35,7 +35,7 @@ export function CommentCard({ item }: CommentCardProps) {
               <UserResourceLink data={item.authorUser} withTooltip hideAvatar />
             </h2>
             <span className="text-xs text-gray-500">
-              {formatToTwinfaceDate(item.createdAt)}
+              {formatToTwinfaceDate(item.createdAt!)}
             </span>
           </div>
