+++ src/entities/comment/libs/hooks/use-filters.ts
@@ -1,11 +1,13 @@
+import { z } from "zod";
+
 import { AutoFormValueInfo, AutoFormValueType } from "@/components/auto-field";
+
 import {
   CommentFilterKeys,
   CommentFilters,
   useCommentSelectAdapter,
 } from "@/entities/comment";
 import { type FilterFeature } from "@/shared/libs";
-import { z } from "zod";
 
 export function useCommentFilters(): FilterFeature<
   CommentFilterKeys,
