+++ src/entities/comment/libs/hooks/use-select-adapter.ts
